# api/docs/swagger.yaml
definitions:
  genproto.CommentsCreateReq:
    properties:
      content:
        type: string
      memory_id:
        type: string
      user_id:
        type: string
    type: object
  genproto.CommentsGetAllRes:
    properties:
      comments:
        items:
          $ref: '#/definitions/genproto.CommentsRes'
        type: array
      count:
        type: integer
    type: object
  genproto.CommentsGetByIdRes:
    properties:
      comment:
        $ref: '#/definitions/genproto.CommentsRes'
    type: object
  genproto.CommentsRes:
    properties:
      content:
        type: string
      id:
        type: string
      memory_id:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.CustomEventsCreateReq:
    properties:
      category:
        type: string
      date:
        type: string
      description:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  genproto.CustomEventsGetAllRes:
    properties:
      count:
        type: integer
      events:
        items:
          $ref: '#/definitions/genproto.CustomEventsRes'
        type: array
    type: object
  genproto.CustomEventsGetByIdRes:
    properties:
      event:
        $ref: '#/definitions/genproto.CustomEventsRes'
    type: object
  genproto.CustomEventsRes:
    properties:
      category:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /memory/{id}/comment:
    post:
      consumes:
      - application/json
      description: Create a new comment
      parameters:
      - description: Comment data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/genproto.CommentsCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Comment created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create comment
      tags:
      - comment
  /memory/{id}/comment/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a comment by ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comment deleted
          schema:
            type: string
        "400":
          description: Invalid comment ID
          schema:
            type: string
        "404":
          description: Comment not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete comment
      tags:
      - comment
    get:
      consumes:
      - application/json
      description: Get a comment
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.CommentsGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get comment
      tags:
      - comment
    put:
      consumes:
      - application/json
      description: Update an existing comment
      parameters:
      - description: Id
        in: query
        name: id
        type: string
      - description: Content
        in: query
        name: content
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comment updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Comment not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update comment
      tags:
      - comment
  /memory/{id}/comment/all:
    get:
      consumes:
      - application/json
      description: Get all comment
      parameters:
      - description: UserId
        in: query
        name: user_id
        type: string
      - description: MemoryId
        in: query
        name: memory_id
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.CommentsGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all comment
      tags:
      - comment
...
  /timeline/custom-event:
    post:
      consumes:
      - application/json
      description: Create a new event
      parameters:
      - description: Event data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/genproto.CustomEventsCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Event created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create event
      tags:
      - event
  /timeline/custom-event/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a event by ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event deleted
          schema:
            type: string
        "400":
          description: Invalid event ID
          schema:
            type: string
        "404":
          description: Event not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete event
      tags:
      - event
    get:
      consumes:
      - application/json
      description: Get a event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.CustomEventsGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get event
      tags:
      - event
    put:
      consumes:
      - application/json
      description: Update an existing event
      parameters:
      - description: Id
        in: query
        name: id
        type: string
      - description: Title
        in: query
        name: title
        type: string
      - description: Description
        in: query
        name: description
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Event not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update event
      tags:
      - event
  /timeline/custom-event/all:
    get:
      consumes:
      - application/json
      description: Get all event
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.CustomEventsGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all event
      tags:
      - event
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"